+++ src/main/java/com/example/demo2/mapper/CommentMapper.java
@@ -1,6 +1,6 @@
 package com.example.demo2.mapper;
 
-import com.example.demo1.board.dto.CommentDto;
+import com.example.demo2.model.dto.CommentDto;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
 
@@ -9,7 +9,7 @@ import java.util.List;
 @Mapper
 @Repository("com.example.demo1.board.mapper.CommentMapper")
 public interface CommentMapper {
-    List<CommentDto> commentList(int id);
+    List<CommentDto> commentList(Long id);
     int commentInsert(CommentDto comment);
     int commentUpdate(CommentDto comment);
     int commentDelete(int id);
